[Front]  IndicatorsRightBar to tsx pure function (#2951)
